fix: bail on first error
Just saw:

    [error]   Cannot replace symbolic link "/root/.zshenv.d" with directory without 'force'
    ln: '/workspaces/.codespaces/.persistedshare/dotfiles/.zshenv.d/locale.zsh' and '/root/.zshenv.d/locale.zsh' are the same file

ie. Fig is trying to create a directory:

    await file({path: '~/.zshenv.d', state: 'directory'});

but it refuses to created a directory because we already have a symlink
here:

    .zshenv.d -> /workspaces/.codespaces/.persistedshare/dotfiles/.zshenv.d/

Even though Fig bails, we then try to call `ln` from our
`codespaces-post-start` script.

So, I'm going to add `force: true` to the `file()` call, but I also want
to bail here early if anything goes wrong.