feat: add bin/up script to make common update operation quicker
Pulls the submodule and creates a commit with the first commit as the
subject line and the full `git submodule summary` output below that.

Leaves you in an editor so you can massage if needed, or `:cq` to abort.

`awk` used to get just second line of `git submodule summary` output,
and `cut` to trim off the leading "  > "; ie. we take:

```
* wincent cb5df9e7...a0f9f1da (1):
  > fix(vim): allow `rollup --watch` to work by setting `backupcopy=yes`
```

and produce just:

```
fix(vim): allow `rollup --watch` to work by setting `backupcopy=yes`
```

Tested by preparing the parent commit.